requirements.txt
Flask-SQLAlchemy==2.5.1
Flask-Migrate==2.7.0
python-dotenv==0.20.0
plivo==4.24.1